Split out new octree_* files
